+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [browser] feat: Show dropped event url in `blacklistUrl`/`whitelistUrl` debug mode
 - [browser] feat: Use better event description instead of `event_id` for user-facing logs
 - [core] ref: Create common integrations that are exposed on `@sentry/core` and reexposed through `browser`/`node`
+- [core] feat: Debug integration
 
 ## 4.0.0